Add subscribe/unsubscribe gauge metric
